+++ modules/avahi.nix
@@ -1,0 +1,8 @@
+{
+  nixosModule = { ... }: {
+    services.avahi.enable = true;
+    services.avahi.publish.enable = true;
+    services.avahi.publish.userServices = true;
+    services.avahi.nssmdns = true;
+  };
+}