+++ modules/base.nix
@@ -68,15 +68,6 @@ in {
       path = "/home/${user}/.zshrc.secrets";
       owner = user;
     };
-
-    # Taken directly from:
-    # https://github.com/NixOS/nixpkgs/blob/HEAD/nixos/modules/services/networking/shairport-sync.nix#L74-L93
-    services.avahi.enable = true;
-    services.avahi.publish.enable = true;
-    services.avahi.publish.userServices = true;
-
-    networking.firewall.allowedTCPPorts = [ 5000 ];
-    networking.firewall.allowedUDPPortRanges = [ { from = 6001; to = 6011; } ];
   };
 
   darwinModule = { user, config, ... }: {