+++ modules/ios.nix
@@ -1,4 +1,6 @@
 {
+  imports = [ "avahi" ];
+
   nixosModule = { pkgs, ... }: {
     environment.systemPackages = builtins.attrValues {
       inherit (pkgs) libimobiledevice;
@@ -6,6 +8,11 @@
 
     # For connecting to iOS devices
     services.usbmuxd.enable = true;
+
+    # Taken directly from:
+    # https://github.com/NixOS/nixpkgs/blob/HEAD/nixos/modules/services/networking/shairport-sync.nix#L74-L93
+    networking.firewall.allowedTCPPorts = [ 5000 ];
+    networking.firewall.allowedUDPPortRanges = [ { from = 6001; to = 6011; } ];
   };
 
   hmModule = { pkgs, lib, ... }: lib.mkIf (pkgs.stdenv.hostPlatform.isLinux) {